Move single choice options one directory up
